# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Visual Reasoning Maps",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks/{task_id}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/docs/{doc_id}": {
      "get": {
        "summary": "Document",
        "operationId": "document_api_docs__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{doc_id}/flow": {
      "get": {
        "summary": "Flow",
        "operationId": "flow_api_docs__doc_id__flow_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{doc_id}/sentences": {
      "get": {
        "summary": "Sentences",
        "operationId": "sentences_api_docs__doc_id__sentences_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sentence"
                  },
                  "title": "Response Sentences Api Docs  Doc Id  Sentences Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{doc_id}/suggestions": {
      "get": {
        "summary": "Suggestions",
        "operationId": "suggestions_api_docs__doc_id__suggestions_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Suggestion"
                  },
                  "title": "Response Suggestions Api Docs  Doc Id  Suggestions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{doc_id}/source": {
      "get": {
        "summary": "Source",
        "operationId": "source_api_docs__doc_id__source_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{doc_id}/assets/{name}": {
      "get": {
        "summary": "Asset",
        "operationId": "asset_api_docs__doc_id__assets__name__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{doc_id}/steps/{step_id}/explanation": {
      "post": {
        "summary": "Explanation",
        "operationId": "explanation_api_docs__doc_id__steps__step_id__explanation_post",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Step Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Explanation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Document": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "enum": [
              "pdf",
              "html"
            ],
            "title": "Kind"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Sections"
          },
          "paragraphs": {
            "items": {
              "$ref": "#/components/schemas/Paragraph"
            },
            "type": "array",
            "title": "Paragraphs"
          },
          "sentences": {
            "items": {
              "$ref": "#/components/schemas/Sentence"
            },
            "type": "array",
            "title": "Sentences"
          },
          "pages": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Pages"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "kind"
        ],
        "title": "Document"
      },
      "Explanation": {
        "properties": {
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "anchors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Anchors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "explanation",
          "anchors"
        ],
        "title": "Explanation"
      },
      "Flow": {
        "properties": {
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "title": "Steps"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "array",
            "title": "Links"
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/TermCard"
            },
            "type": "array",
            "title": "Terms"
          },
          "genre": {
            "type": "string",
            "title": "Genre",
            "default": "scientific"
          },
          "thesis": {
            "type": "string",
            "title": "Thesis",
            "default": ""
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata"
          },
          "patterns": {
            "items": {
              "$ref": "#/components/schemas/Pattern"
            },
            "type": "array",
            "title": "Patterns"
          },
          "generation": {
            "$ref": "#/components/schemas/Generation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "steps",
          "links",
          "metadata",
          "generation"
        ],
        "title": "Flow"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Link": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "src": {
            "type": "string",
            "title": "Src"
          },
          "dst": {
            "type": "string",
            "title": "Dst"
          },
          "type": {
            "type": "string",
            "enum": [
              "support",
              "cause",
              "refine",
              "contradict"
            ],
            "title": "Type"
          },
          "connective": {
            "type": "string",
            "title": "Connective",
            "default": ""
          },
          "anchors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Anchors"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "enum": [
              "verified",
              "partial",
              "to_verify"
            ],
            "title": "Status",
            "default": "to_verify"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "src",
          "dst",
          "type",
          "anchors",
          "confidence"
        ],
        "title": "Link"
      },
...
      "Paragraph": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "words": {
            "items": {
              "$ref": "#/components/schemas/Word"
            },
            "type": "array",
            "title": "Words"
          },
          "dom_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dom Id"
          },
          "latex": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Latex"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text",
          "section_id"
        ],
        "title": "Paragraph"
      },
...
      "Section": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "Section"
      },
      "Sentence": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "bbox": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 4,
              "minItems": 4
            },
            "type": "array",
            "title": "Bbox"
          },
          "dom_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dom Id"
          },
          "paragraph_id": {
            "type": "string",
            "title": "Paragraph Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text",
          "section_id",
          "order",
          "paragraph_id"
        ],
        "title": "Sentence"
      },
      "Step": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "type": {
            "type": "string",
            "enum": [
              "question",
              "premise",
              "claim",
              "evidence",
              "objection",
              "conclusion"
            ],
            "title": "Type"
          },
          "label": {
            "type": "string",
            "maxLength": 40,
            "title": "Label"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "anchors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Anchors"
          },
          "quote": {
            "type": "string",
            "minLength": 1,
            "title": "Quote"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "enum": [
              "verified",
              "partial",
              "to_verify"
            ],
            "title": "Status",
            "default": "to_verify"
          },
          "first_position": {
            "type": "integer",
            "title": "First Position",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "label",
          "summary",
          "anchors",
          "quote",
          "confidence"
        ],
        "title": "Step"
      },
      "Suggestion": {
        "properties": {
          "target_type": {
            "type": "string",
            "enum": [
              "step",
              "link"
            ],
            "title": "Target Type"
          },
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Severity"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "suggested_rewrite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Rewrite"
          },
          "anchors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Anchors"
          },
          "source": {
            "type": "string",
            "enum": [
              "rule",
              "llm"
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_type",
          "target_id",
          "category",
          "severity",
          "message",
          "anchors",
          "source"
        ],
        "title": "Suggestion"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}